+++ test/html_pipeline/text_filter/markdown_filter_test.rb
@@ -1,0 +1,180 @@
+# frozen_string_literal: true
+
+require "test_helper"
+
+MarkdownFilter = HTMLPipeline::TextFilter::MarkdownFilter
+
+class HTMLPipeline
+  class MarkdownFilterTest < Minitest::Test
+    class CustomRenderer < CommonMarker::HtmlRenderer
+      def header(node)
+        block do
+          text = node.first_child.string_content
+          level = node.header_level
+          out("{level: #{level}, text: #{text}}")
+          super
+        end
+      end
+    end
+
+    def setup
+      @haiku =
+        "Pointing at the moon\n" \
+          "Reminded of simple things\n" \
+          "Moments matter most"
+      @links =
+        "See http://example.org/ for more info"
+      @code =
+        "```\n" \
+          "def hello()" \
+          "  'world'" \
+          "end" \
+          "```"
+      @header = <<~DOC
+        # Words
+
+        Some words
+
+        ## Words
+
+        More words?
+      DOC
+    end
+
+    def test_fails_when_given_a_documentfragment
+      body = "<p>heyo</p>"
+      doc  = HTMLPipeline.parse(body)
+      assert_raises(TypeError) { MarkdownFilter.call(doc, context: {}) }
+    end
+
+    def test_gfm_enabled_by_default
+      doc = MarkdownFilter.to_document(@haiku)
+      assert(doc.is_a?(HTMLPipeline::DocumentFragment))
+      assert_equal(2, doc.search("br").size)
+    end
+
+    def test_disabling_gfm
+      doc = MarkdownFilter.to_document(@haiku, context: { gfm: false })
+      assert(doc.is_a?(HTMLPipeline::DocumentFragment))
+      assert_equal(0, doc.search("br").size)
+    end
+
+    def test_fenced_code_blocks
+      doc = MarkdownFilter.to_document(@code)
+      assert(doc.is_a?(HTMLPipeline::DocumentFragment))
+      assert_equal(1, doc.search("pre").size)
+    end
+
+    def test_fenced_code_blocks_with_language
+      doc = MarkdownFilter.to_document(@code.sub("```", "``` ruby"))
+      assert(doc.is_a?(HTMLPipeline::DocumentFragment))
+      assert_equal(1, doc.search("pre").size)
+      assert_equal("ruby", doc.search("pre").first["lang"])
+    end
+
+    def test_standard_extensions
+      iframe = "<iframe src='http://www.google.com'></iframe>"
+      iframe_escaped = "&lt;iframe src='http://www.google.com'>&lt;/iframe>"
+      doc = MarkdownFilter.new(iframe, context: { unsafe: true }).call
+      assert_equal(doc, iframe_escaped)
+    end
+
+    def test_changing_extensions
+      iframe = "<iframe src='http://www.google.com'></iframe>"
+      doc = MarkdownFilter.new(iframe, context: { commonmarker_extensions: [], unsafe: true }).call
+      assert_equal(doc, iframe)
+    end
+
+    def test_bogus_renderer
+      assert_raises(ArgumentError) do
+        MarkdownFilter.to_document(@haiku, context: { commonmarker_renderer: 23 })
+      end
+    end
+
+    def test_legitimate_renderer
+      results = MarkdownFilter.new(@header, context: { commonmarker_renderer: CustomRenderer }).call
+      expected = <<~DOC
+        {level: 1, text: Words}
+        <h1>Words</h1>
+        <p>Some words</p>
+        {level: 2, text: Words}
+        <h2>Words</h2>
+        <p>More words?</p>
+      DOC
+
+      assert_equal(results, expected.chomp)
+    end
+
+    def test_without_tagfilter
+      extensions = HTMLPipeline::TextFilter::MarkdownFilter::DEFAULT_COMMONMARKER_EXTENSIONS - [:tagfilter]
+      script = "<script>foobar</script>"
+      results = MarkdownFilter.new(script, context: { unsafe: true, commonmarker_extensions: extensions }).call
+
+      assert_equal(results, script)
+    end
+
+    def test_legitimate_custom_renderer_without_tagfilter
+      extensions = HTMLPipeline::TextFilter::MarkdownFilter::DEFAULT_COMMONMARKER_EXTENSIONS - [:tagfilter]
+      script = "<script>foobar</script>"
+      results = MarkdownFilter.new(script, context: { unsafe: true, commonmarker_extensions: extensions, commonmarker_renderer: CustomRenderer }).call
+
+      assert_equal(results, script)
+    end
+
+    def test_unsafe_custom_renderer_fenced_code_blocks_with_language
+      doc = MarkdownFilter.to_document(@code.sub("```", "``` ruby"), context: { unsafe: true, commonmarker_renderer: CustomRenderer })
+      assert_equal("ruby", doc.search("pre").first["lang"])
+    end
+  end
+end
+
+class GFMTest < Minitest::Test
+  def setup
+    @gfm = MarkdownFilter
+    @context = { gfm: true, unsafe: true }
+  end
+
+  def test_not_touch_single_underscores_inside_words
+    assert_equal("<p>foo_bar</p>",
+      @gfm.call("foo_bar", context: @context))
+  end
+
+  def test_not_touch_underscores_in_code_blocks
+    assert_equal("<pre><code>foo_bar_baz\n</code></pre>",
+      @gfm.call("    foo_bar_baz", context: @context))
+  end
+
+  def test_not_touch_underscores_in_pre_blocks
+    assert_equal("<pre>\nfoo_bar_baz\n</pre>",
+      @gfm.call("<pre>\nfoo_bar_baz\n</pre>", context: @context))
+  end
+
+  def test_not_touch_two_or_more_underscores_inside_words
+    assert_equal("<p>foo_bar_baz</p>",
+      @gfm.call("foo_bar_baz", context: @context))
+  end
+
+  def test_turn_newlines_into_br_tags_in_simple_cases
+    assert_equal("<p>foo<br />\nbar</p>",
+      @gfm.call("foo\nbar", context: @context))
+  end
+
+  def test_convert_newlines_in_all_groups
+    assert_equal("<p>apple<br />\npear<br />\norange</p>\n" \
+      "<p>ruby<br />\npython<br />\nerlang</p>",
+      @gfm.call("apple\npear\norange\n\nruby\npython\nerlang", context: @context))
+  end
+
+  def test_convert_newlines_in_even_long_groups
+    assert_equal("<p>apple<br />\npear<br />\norange<br />\nbanana</p>\n" \
+      "<p>ruby<br />\npython<br />\nerlang</p>",
+      @gfm.call("apple\npear\norange\nbanana\n\nruby\npython\nerlang", context: @context))
+  end
+
+  def test_not_convert_newlines_in_lists
+    assert_equal("<h1>foo</h1>\n<h1>bar</h1>",
+      @gfm.call("# foo\n# bar", context: @context))
+    assert_equal("<ul>\n<li>foo</li>\n<li>bar</li>\n</ul>",
+      @gfm.call("* foo\n* bar", context: @context))
+  end
+end